# feat: support Set equality comparison in iterableEquals