--- a/lib/src/equatable.dart
+++ b/lib/src/equatable.dart
@@ -46,7 +46,7 @@ abstract class Equatable {
     return identical(this, other) ||
         other is Equatable &&
             runtimeType == other.runtimeType &&
-            iterableEquals(props, other.props);
+            propsEquals(props, other.props);
   }
 
   @override