@ShakyaCsun 2025-04-17
1. the `setEquals` function can be safely deleted now and replaced with `iterableEquals` in the tests but looking at the comment on `equals` function i.e. #187, I decided to leave it in for now.
1. I used `propsEquals` instead of modifying and using `equals` to avoid checking for `null`, but maybe we can remove nullable params from `equals` instead?
1. I also noticed we could change `mapPropsToHashCode` to take `List<Object?>` instead of `Iterable<Object?>?` and avoid null checks there as well but wanted to get your thoughts before changing function signatures even though these utils aren't publicly exported.
1. Even though Equatable objects with Iterable and List with same elements are no longer equal, they still produce the same hashCode. One idea I had was to change https://github.com/felangel/equatable/blob/834d16779c9eb36b1bc769a99a6f16f6605a9bb6/lib/src/equatable_utils.dart#L102 to 
```dart
return hash ^ (object is List ? 'List${object.length}'.hashCode : object.length);
```
but maybe there's a better way